Crate axum_macros

Macros for axum.

Attribute Macros

debug_handler

Generates better error messages when applied to handler functions.

debug_middleware

Generates better error messages when applied to middleware functions.

Derive Macros

FromRef

Derive an implementation of FromRef for each field in a struct.

FromRequest

Derive an implementation of FromRequest.

FromRequestParts

Derive an implementation of FromRequestParts.

TypedPath

Derive an implementation of axum_extra::routing::TypedPath.